Refactor server to own folder
