+++ server/server.js
@@ -9,42 +9,13 @@ import { StaticRouter } from 'react-router-dom';
 import StyleContext from 'isomorphic-style-loader/StyleContext';
 import thunk from 'redux-thunk';
 import config from '../config';
-import rootReducer from './rootReducer';
-import App from './App';
-import I18n from './i18n';
+import rootReducer from '../src/rootReducer';
+import App from '../src/App';
+import {handleRequestRedirect} from './utils';
 
 // Configure constants
 const app = express();
 const serverConfig = config.server;
-const allowedUrls = [
-  /^\/.{2,}\/$/,
-  /^\/.{2,}\/unit\/\d+$/,
-  /^\/.{2,}\/unit$/,
-  /^\/.{2,}\/search$/,
-  /^\/.{2,}\/address\/[^\/]+\/[^\/]+\/[^\/]+$/,
-  /^\/.{2,}\/division\/[^\/]+\/[^\/]+$/,
-  /^\/.{2,}\/division$/,
-  /^\/.{2,}\/area$/,
-];
-
-const handleRequestRedirect = (req, res) => {
-  // Check if request url is actual path
-  let match = false;
-  allowedUrls.forEach((url) => {
-    if (!match && req.path.match(url)) {
-      match = true;
-    }
-  });
-
-  // Handle language check and redirect if language is changed to default
-  const i18n = new I18n();
-  const pathArray = req.url.split('/');
-  if (match && !i18n.isValidLocale(pathArray[1])) {
-    pathArray[1] = i18n.locale;
-    res.redirect(pathArray.join('/'));
-    return true;
-  }
-};
 
 app.use(express.static(path.resolve(__dirname, 'src')));
